+++ tests/model/test_amortizedlda.py
@@ -1,12 +1,14 @@
 import os
 
 import numpy as np
+import pytest
 
 from scvi.data import synthetic_iid
 from scvi.model import AmortizedLDA
 
 
-def test_lda_model_single_step(n_topics: int = 5):
+@pytest.mark.parametrize("n_topics", [5])
+def test_lda_model_single_step(n_topics: int):
     adata = synthetic_iid()
     AmortizedLDA.setup_anndata(adata)
     mod1 = AmortizedLDA(adata, n_topics=n_topics, cell_topic_prior=1.5, topic_feature_prior=1.5)
@@ -14,7 +16,8 @@ def test_lda_model_single_step(n_topics: int = 5):
     assert len(mod1.history["elbo_train"]) == 1
 
 
-def test_lda_model_single_step_with_external_indices(n_topics: int = 5):
+@pytest.mark.parametrize("n_topics", [5])
+def test_lda_model_single_step_with_external_indices(n_topics: int):
     adata = synthetic_iid()
     AmortizedLDA.setup_anndata(adata)
     mod1 = AmortizedLDA(adata, n_topics=n_topics, cell_topic_prior=1.5, topic_feature_prior=1.5)
@@ -37,7 +40,8 @@ def test_lda_model_single_step_with_external_indices(n_topics: int = 5):
     assert len(mod1.history["elbo_train"]) == 1
 
 
-def test_lda_model(n_topics: int = 5):
+@pytest.mark.parametrize("n_topics", [5])
+def test_lda_model(n_topics: int):
     adata = synthetic_iid()
 
     # Test with float and Sequence priors.
@@ -85,7 +89,8 @@ def test_lda_model(n_topics: int = 5):
     mod.get_perplexity(adata2)
 
 
-def test_lda_model_save_load(save_path: str, n_topics: int = 5):
+@pytest.mark.parametrize("n_topics", [5])
+def test_lda_model_save_load(save_path: str, n_topics: int):
     adata = synthetic_iid()
     AmortizedLDA.setup_anndata(adata)
     mod = AmortizedLDA(adata, n_topics=n_topics)