+++ tests/autotune/test_tune.py
@@ -43,7 +43,8 @@ def test_run_autotune_scvi_basic(save_path: str):
 
 
 @pytest.mark.autotune
-def test_run_autotune_scvi_no_anndata(save_path: str, n_batches: int = 3):
+@pytest.mark.parametrize("n_batches", [3])
+def test_run_autotune_scvi_no_anndata(save_path: str, n_batches: int):
     from ray import tune
     from ray.tune import ResultGrid
 
@@ -150,7 +151,8 @@ def test_run_autotune_scvi_with_scib(model_cls, metric: str, save_path: str):
 
 
 @pytest.mark.autotune
-def test_run_autotune_scvi_with_scib_ext_indices(save_path: str, metric: str = "iLISI"):
+@pytest.mark.parametrize("metric", ["iLISI"])
+def test_run_autotune_scvi_with_scib_ext_indices(save_path: str, metric: str):
     from ray import tune
     from ray.tune import ResultGrid