+++ CHANGELOG.md
@@ -34,6 +34,8 @@ to [Semantic Versioning]. Full commit history is available in the
 
 #### Changed
 
+- Update Read the docs tutorials with one main preprocessing tutorial {pr}`3363`.
+
 #### Removed
 
 ### 1.3.1 (2025-05-15)